[core] Move test to test_serialization.py, use two-level nested func
