--- python/tests/BUILD
@@ -1,10 +1,0 @@
-load("@rules_python//python:defs.bzl", "py_test")
-
-py_test(
-    name = "test_check_serialize",
-    size = "small",
-    srcs = ["test_check_serialize.py"],
-    deps = [
-        "//python/ray/util:check_serialize",
-    ],
-)
